requirements.txt
pandas
python-dotenv
spotipy
SQLAlchemy
psycopg2-binary
pytz